requirements.txt
streamlit==1.34.0
tensorflow==2.15.0
Pillow
numpy
pandas
googletrans==4.0.0-rc1
deep-translator
